# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ForumCategory:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      order:
        type: integer
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  models.ForumPost:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_edited:
        type: boolean
      topic_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.ForumTopic:
    properties:
      category_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_locked:
        type: boolean
      is_pinned:
        type: boolean
      last_post_at:
        type: string
      tenant_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      views:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A chat system API with real-time messaging, file sharing, and support
    ticket management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Chat System API
  version: "1.0"
paths:
...
  /api/v1/forum/categories:
    get:
      consumes:
      - application/json
      description: Get all forum categories
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            items:
              $ref: '#/definitions/models.ForumCategory'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Get forum categories
      tags:
      - Forum
    post:
      consumes:
      - application/json
      description: Create a new forum category
      parameters:
      - description: Category details
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.ForumCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            $ref: '#/definitions/models.ForumCategory'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Create forum category
      tags:
      - Forum
  /api/v1/forum/categories/{id}/topics:
    get:
      consumes:
      - application/json
      description: Get all topics in a specific category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of topics per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of topics
          schema:
            items:
              $ref: '#/definitions/models.ForumTopic'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties: true
            type: object
      summary: Get topics in category
      tags:
      - Forum
  /api/v1/forum/posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing forum post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Post deleted successfully
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
      summary: Delete forum post
      tags:
      - Forum
    put:
      consumes:
      - application/json
      description: Update an existing forum post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated post details
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.ForumPost'
      produces:
      - application/json
      responses:
        "200":
          description: Post updated successfully
          schema:
            $ref: '#/definitions/models.ForumPost'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
      summary: Update forum post
      tags:
      - Forum
  /api/v1/forum/topics:
    post:
      consumes:
      - application/json
      description: Create a new forum topic
      parameters:
      - description: Topic details
        in: body
        name: topic
        required: true
        schema:
          $ref: '#/definitions/models.ForumTopic'
      produces:
      - application/json
      responses:
        "201":
          description: Topic created successfully
          schema:
            $ref: '#/definitions/models.ForumTopic'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Create forum topic
      tags:
      - Forum
  /api/v1/forum/topics/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing forum topic
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Topic deleted successfully
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Topic not found
          schema:
            additionalProperties: true
            type: object
      summary: Delete forum topic
      tags:
      - Forum
    put:
      consumes:
      - application/json
      description: Update an existing forum topic
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated topic details
        in: body
        name: topic
        required: true
        schema:
          $ref: '#/definitions/models.ForumTopic'
      produces:
      - application/json
      responses:
        "200":
          description: Topic updated successfully
          schema:
            $ref: '#/definitions/models.ForumTopic'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Topic not found
          schema:
            additionalProperties: true
            type: object
      summary: Update forum topic
      tags:
      - Forum
  /api/v1/forum/topics/{id}/posts:
    get:
      consumes:
      - application/json
      description: Get all posts in a forum topic
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of posts per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of posts
          schema:
            items:
              $ref: '#/definitions/models.ForumPost'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Topic not found
          schema:
            additionalProperties: true
            type: object
      summary: Get forum posts
      tags:
      - Forum
    post:
      consumes:
      - application/json
      description: Create a new post in a forum topic
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      - description: Post details
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.ForumPost'
      produces:
      - application/json
      responses:
        "201":
          description: Post created successfully
          schema:
            $ref: '#/definitions/models.ForumPost'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Topic not found
          schema:
            additionalProperties: true
            type: object
      summary: Create forum post
      tags:
      - Forum
  /api/v1/forum/topics/{id}/subscribe:
    delete:
      consumes:
      - application/json
      description: Unsubscribe from notifications for a forum topic
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Unsubscribed successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Topic not found
          schema:
            additionalProperties: true
            type: object
      summary: Unsubscribe from topic
      tags:
      - Forum
    post:
      consumes:
      - application/json
      description: Subscribe to notifications for a forum topic
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Subscribed successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Topic not found
          schema:
            additionalProperties: true
            type: object
      summary: Subscribe to topic
      tags:
      - Forum
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...